+++ crates/migrate-nx/CHANGELOG.md
@@ -1,3 +1,13 @@
+# Changelog
+
+## Unreleased
+
+#### 🚀 Updates
+
+- Added support for `defaultBase` from `nx.json`.
+- Added support for `metadata` from `project.json`.
+- Updated dependencies.
+
 ## 0.0.5
 
 #### ⚙️ Internal